+++ Tweakr.php
@@ -1,16 +1,18 @@
 <?php
 /**
-    Plugin Name: Tweakr Toolkit
-    Plugin URI: https://andidittrich.de/go/tweakr
-    Description: Supercharges your Blog with common used Tweaks, Features and Utilities
-    Version: 1.4
-    Author: Andi Dittrich
+    Plugin Name: Tweakr - Utility Toolkit
+    Plugin URI: https://github.com/AndiDittrich/WordPress.Tweakr
+    Description: Supercharges your Blog with WP Core Tweaks, Advanced Features and Utilities
+    Version: 1.5-BETA1
+    Author: Andi Dittrich, Aenon Dynamics
     Author URI: https://andidittrich.de
     License: MIT X11 License
+    Text Domain: tweakr
+    Domain Path: /lang
 
     ----
     The MIT License (X11 License)
-    Copyright (c) 2016-2017 Andi Dittrich <https://andidittrich.de>
+    Copyright (c) 2016-2018 Andi Dittrich <https://andidittrich.de>
     Permission is hereby granted, free of charge, to any personobtaining a copy of this software and associated documentationfiles (the "Software"), to deal in the Software withoutrestriction, including without limitation the rights to use,copy, modify, merge, publish, distribute, sublicense, and/or sellcopies of the Software, and to permit persons to whom theSoftware is furnished to do so, subject to the followingconditions:
     The above copyright notice and this permission notice shall beincluded in all copies or substantial portions of the Software.
     THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIESOF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE ANDNONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHTHOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISINGFROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OROTHER DEALINGS IN THE SOFTWARE.
@@ -23,7 +25,9 @@
     Plugin Bootstrap Operation
 */
 define('TWEAKR_INIT', true);
-define('TWEAKR_VERSION', '1.4');
+define('TWEAKR_VERSION', '1.5-BETA1');
+define('TWEAKR_PLUGIN_TITLE', 'Tweakr - Utility Toolkit');
+define('TWEAKR_PLUGIN_HEADLINE', 'Supercharges your Blog with WP Core Tweaks, Advanced Features and Utilities');
 define('TWEAKR_PLUGIN_PATH', dirname(__FILE__));
 define('TWEAKR_PLUGIN_URL', plugins_url('/tweakr/'));
 
@@ -57,10 +61,12 @@ if (version_compare(phpversion(), '5.4', '>=')){
     require_once(TWEAKR_PLUGIN_PATH.'/class/GoogleAnalytics.php');
     require_once(TWEAKR_PLUGIN_PATH.'/class/HttpHeader.php');
     require_once(TWEAKR_PLUGIN_PATH.'/class/LinkManager.php');
+    require_once(TWEAKR_PLUGIN_PATH.'/class/Metadata.php');
     require_once(TWEAKR_PLUGIN_PATH.'/class/Monitoring.php');
     require_once(TWEAKR_PLUGIN_PATH.'/class/PiwikAnalytics.php');
     require_once(TWEAKR_PLUGIN_PATH.'/class/ResourceLoader.php');
     require_once(TWEAKR_PLUGIN_PATH.'/class/RewriteRules.php');
+    require_once(TWEAKR_PLUGIN_PATH.'/class/Robots.php');
     require_once(TWEAKR_PLUGIN_PATH.'/class/TinyMCE.php');
     require_once(TWEAKR_PLUGIN_PATH.'/class/Tweakr.php');
     require_once(TWEAKR_PLUGIN_PATH.'/class/UserNotification.php');